moved setComponentHierarchy
